requirements.txt
nltk
sentence-transformers
requests
python-dotenv
pypdf
openai
openai-agents
pypdf2
python-docx
numpy
chromadb
